// PropertiesScraper/src/models/zillowHouse.ts
interface ZillowHouse {
    "zpid": string;
    "id": string;
    "providerListingId": any;
    "imgSrc": string;
    "hasImage"?: boolean;
    "detailUrl": string;
    "statusType": string;
    "statusText": string;
    "countryCurrency": string;
    "price": number;
    "address": string;
    "addressStreet": string;
    "addressCity": string;
    "addressState": string;
    "addressZipcode": string;
    "isUndisclosedAddress": boolean;
    "lotAreaString"?: string;
    "beds": any;
    "baths": any;
    "area": any;
    "latitude": number;
    "longitude": number;
    "isZillowOwned": boolean;
    "variableData": any;
    "badgeInfo": any;
    "hdpData": {
        "homeInfo": {
            "zpid": number;
            "streetAddress": string;
            "zipcode": string;
            "city": string;
            "state": string;
            "latitude": number;
            "longitude": number;
            "price": number;
            "datePriceChanged"?: number;
            "homeType": string;
            "homeStatus": string;
            "daysOnZillow": number;
            "isFeatured": boolean;
            "shouldHighlight": boolean;
            "listing_sub_type": any;
            "priceReduction"?: string;
            "isUnmappable": boolean;
            "isPreforeclosureAuction": boolean;
            "homeStatusForHDP": string;
            "priceForHDP": number
            "priceChange"?: number;
            "isNonOwnerOccupied": boolean;
            "isPremierBuilder": boolean;
            "isZillowOwned": boolean;
            "currency": string;
            "country": string;
            "unit"?: string;
            "lotAreaValue": number;
            "lotAreaUnit": string;
            "livingArea": number;
            "info3String"?: string,
            "info2String"?: string,
            "info1String"?: string,
            "bathrooms"?: 2,
            "bedrooms"?: 3,
            "taxAssessedValue"?: 324640,

        }
    };
    "isSaved": boolean;
    "isUserClaimingOwner": boolean;
    "isUserConfirmedClaim": boolean;
    "pgapt": string;
    "sgapt": string;
    "zestimate": any;
    "shouldShowZestimateAsPrice": boolean;
    "has3DModel": boolean;
    "hasVideo": boolean;
    "isHomeRec": boolean;
    "info1String"?: string,
    "info2String"?: string;
    "info3String"?: string;
    "brokerName"?: string;
    "hasAdditionalAttributions": boolean;
    "isFeaturedListing": boolean;
    "availabilityDate": any;
    "list": boolean;
    "relaxed": boolean;
    "maxPagination": number;
    "unformattedPrice": number,
    "latLong"?: { "latitude": number, "longitude": number },

}
export default ZillowHouse